add homepage/online/offline functions
