+++ requirements.txt
@@ -1,4 +1,5 @@
 attrs==20.3.0
+aiofiles
 certifi==2020.12.5
 cffi==1.14.5
 chardet==4.0.0
