+++ templates/components/card.html
@@ -1,6 +1,6 @@
 <div class="card" style="width: 18rem;">
   <div class="card-body">
-    <img src="{{ url_for('static', filename = '001-youtube.png') }}" width="16" height="16" alt="previous" title="Previous" border="0">
+    <img src="{{ url_for('static', path='images/001-youtube.png') }}" width="16" height="16" alt="previous" title="Previous" border="0">
     <h5 class="card-title">{{ obj.website_name}}</h5>
     <p class="card-text">{{ obj.website_card_desc }}</p>
     <p class="card-text">{{ obj.website_url }}</p>
