+++ templates/components/navbar.html
